backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
requests==2.31.0
python-dateutil==2.8.2
gunicorn==21.2.0
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
